% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_Mplus_file.R
\name{create_Mplus_file}
\alias{create_Mplus_file}
\title{Create Mplus file}
\usage{
create_Mplus_file(x, y, z, path)
}
\arguments{
\item{x}{number of time points}

\item{y}{number of replications}

\item{z}{effect size in the code}

\item{path}{name of the folder to save datasets}
}
\value{

}
\description{
Create Mplus file
}
\examples{
create_Mplus_file(12,50, -0.33, "Power1")
}
