% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_power.R
\name{estimate_power}
\alias{estimate_power}
\title{Estimate Power}
\usage{
estimate_power(path)
}
\arguments{
\item{path}{Name of the folder with your data}
}
\value{

}
\description{
Estimate Power
}
\examples{
estimate_power_create_file("Power1")
}
