% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_Mplus_estimate_power.R
\name{estimate_power_create_file}
\alias{estimate_power_create_file}
\title{Create Mplus Estimate Power}
\usage{
estimate_power_create_file(path)
}
\arguments{
\item{path}{Name of the folder with your data}
}
\value{

}
\description{
Create Mplus Estimate Power
}
\examples{
estimate_power_create_file("Power1")

}
