% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_the_data.R
\name{simulate_the_data}
\alias{simulate_the_data}
\title{Simulate the Data}
\usage{
simulate_the_data(path)
}
\arguments{
\item{path}{Name of the folder with your data}
}
\value{

}
\description{
Simulate the Data
}
\examples{
simulate_the_data("Power1")
}
